Test Merge - Multiple Fixes (#317)
* clean up

* Update runspace.ps1 for new tweaks (#291)

* clean up (#290)

* Update Tweaks


* Github actions to update script based on branch (#294)

* clean up (#290)

* Update Tweaks

* test pipeline

* test pipeline

* chore: autopublish 2022-10-19T09:34:35Z

* test pipeline

* test pipeline

* chore: autopublish 2022-10-19T11:39:50Z

* test pipeline

* test pipeline

* test pipeline

* test pipeline

* test pipeline

* test pipeline

* test

* please work

* Update Branch in script

* test new variable

* Update Branch in script

* test all branches

* updated $BranchToUse to use the correct casing


* bring test up to date (#295)

* clean up (#290)

* Update README.md

* Update README.md

* Update README.md

* Update release.yaml

* Replace Broken Issue Links in KnownIssues.md (#297)

- The issues listed in KnownIssues.md were broken due to the issues tab in the previous repo being removed.
  - Replaced with issues in this repo that apply to the issues, if there were any
- I think this issue could be applied to the main branch without worry, but I'll leave that up to you.

* Update Branch in script

* Implement check for local ooshutup10.cfg file (#298)

- Checks for an ooshutup10.cfg file in the current directory and uses it if found
  - This allows users to use custom config files if they wish

* bring winget install script (#302)

* clean up (#290)

* Update README.md

* Update README.md

* Update README.md

* Update winget.ps1

* Update Branch in script

* Brave Fix

* Fix Flickering

New Windows versions HATE timeout tweaks

* Fixing Error Messages

* syntax fix

* Switch to PSGallery winget-installer

* Adding Branch Variables from #309

+++ winutil.ps1
@@ -1,3 +1,5 @@
+#for CI/CD
+$BranchToUse = 'test'
 <#
 .NOTES
    Author      : Chris Titus @christitustech
@@ -5,8 +7,7 @@
     Version 0.0.1
 #>
 
-#$inputXML = Get-Content "MainWindow.xaml" #uncomment for development
-$inputXML = (new-object Net.WebClient).DownloadString("https://raw.githubusercontent.com/ChrisTitusTech/winutil/main/MainWindow.xaml") #uncomment for Production
+$inputXML = (new-object Net.WebClient).DownloadString("https://raw.githubusercontent.com/ChrisTitusTech/winutil/$BranchToUse/MainWindow.xaml") #uncomment for Production
 
 $inputXML = $inputXML -replace 'mc:Ignorable="d"', '' -replace "x:N", 'N' -replace '^<Win.*', '<Window'
 [void][System.Reflection.Assembly]::LoadWithPartialName('presentationframework')
@@ -132,7 +133,7 @@ $WPFinstall.Add_Click({
             $WPFInstallautohotkey.IsChecked = $false
         }  
         If ( $WPFInstallbrave.IsChecked -eq $true ) { 
-            $wingetinstall.Add("BraveSoftware.BraveBrowser")
+            $wingetinstall.Add("Brave.Brave")
             $WPFInstallbrave.IsChecked = $false
         }
         If ( $WPFInstallchrome.IsChecked -eq $true ) { 
@@ -545,33 +546,25 @@ $WPFinstall.Add_Click({
             Write-Host "Winget Already Installed"
         }
         else {
-            if (((((Get-ComputerInfo).OSName.IndexOf("LTSC")) -ne -1) -or ((Get-ComputerInfo).OSName.IndexOf("Server") -ne -1)) -and (((Get-ComputerInfo).WindowsVersion) -ge "1809")) {
-                #Checks if Windows edition is LTSC/Server 2019+
-                #Manually Installing Winget
+            #Gets the computer's information
+            $ComputerInfo = Get-ComputerInfo
+
+            #Gets the Windows Edition
+            $OSName = if ($ComputerInfo.OSName) {
+                $ComputerInfo.OSName
+            }else {
+                $ComputerInfo.WindowsProductName
+            }
+
+            if (((($OSName.IndexOf("LTSC")) -ne -1) -or ($OSName.IndexOf("Server") -ne -1)) -and (($ComputerInfo.WindowsVersion) -ge "1809")) {
+                
                 Write-Host "Running Alternative Installer for LTSC/Server Editions"
 
-                #Download Needed Files
-                Write-Host "Downloading Needed Files..."
-                Start-BitsTransfer -Source "https://aka.ms/Microsoft.VCLibs.x64.14.00.Desktop.appx" -Destination "$env:TEMP\Microsoft.VCLibs.x64.14.00.Desktop.appx"
-                Start-BitsTransfer -Source "https://github.com/microsoft/winget-cli/releases/download/v1.2.10271/Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle" -Destination "$env:TEMP\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle"
-                Start-BitsTransfer -Source "https://github.com/microsoft/winget-cli/releases/download/v1.2.10271/b0a0692da1034339b76dce1c298a1e42_License1.xml" -Destination "$env:TEMP\b0a0692da1034339b76dce1c298a1e42_License1.xml"
-
-                #Installing Packages
-                Write-Host "Installing Packages..."
-                Add-AppxProvisionedPackage -Online -PackagePath "$env:TEMP\Microsoft.VCLibs.x64.14.00.Desktop.appx" -SkipLicense
-                Add-AppxProvisionedPackage -Online -PackagePath "$env:TEMP\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle" -LicensePath "$env:TEMP\b0a0692da1034339b76dce1c298a1e42_License1.xml"
-                Write-Host "winget Installed (Reboot might be required before winget will work)"
-
-                #Sleep for 5 seconds to maximize chance that winget will work without reboot
-                Write-Host "Pausing for 5 seconds to maximize chance that winget will work without reboot"
-                Start-Sleep -s 5
-
-                #Removing no longer needed Files
-                Write-Host "Removing no longer needed Files..."
-                Remove-Item -Path "$env:TEMP\Microsoft.VCLibs.x64.14.00.Desktop.appx" -Force
-                Remove-Item -Path "$env:TEMP\Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle" -Force
-                Remove-Item -Path "$env:TEMP\b0a0692da1034339b76dce1c298a1e42_License1.xml" -Force
-                Write-Host "Removed Files that are no longer needed"
+                # Switching to winget-install from PSGallery from asheroto
+                # Source: https://github.com/asheroto/winget-installer
+                
+                Start-Process powershell.exe -Verb RunAs -ArgumentList "-command irm https://raw.githubusercontent.com/asheroto/winget-installer/master/winget-install.ps1 | iex | Out-Host" -WindowStyle Normal
+                
             }
             elseif (((Get-ComputerInfo).WindowsVersion) -lt "1809") {
                 #Checks if Windows Version is too old for winget
@@ -755,15 +748,14 @@ $WPFtweaksbutton.Add_Click({
 
         If ( $WPFEssTweaksDeleteTempFiles.IsChecked -eq $true ) {
             Write-Host "Delete Temp Files"
-            Get-ChildItem -Path "C:\Windows\Temp" *.* -Recurse | Remove-Item -Force -Recurse
-            Get-ChildItem -Path $env:TEMP *.* -Recurse | Remove-Item -Force -Recurse
+            Get-ChildItem -Path "C:\Windows\Temp" *.* -Recurse | Remove-Item -Force -Recurse -ErrorAction SilentlyContinue
+            Get-ChildItem -Path $env:TEMP *.* -Recurse | Remove-Item -Force -Recurse -ErrorAction SilentlyContinue
             $WPFEssTweaksDeleteTempFiles.IsChecked = $false
-            Write-Host "================================="
-            Write-Host "--- !!!!ERRORS ARE NORMAL!!!! ---"
-            Write-Host "--- Cleaned following folders:---"
-            Write-Host "--- C:\Windows\Temp           ---"
-            Write-Host "---"$env:TEMP"---"
-            Write-Host "================================="
+            Write-Host "======================================="
+            Write-Host "--- Cleaned following folders:"
+            Write-Host "--- C:\Windows\Temp"
+            Write-Host "--- "$env:TEMP
+            Write-Host "======================================="
         }
 
         If ( $WPFEssTweaksDVR.IsChecked -eq $true ) {
@@ -775,6 +767,9 @@ $WPFtweaksbutton.Add_Click({
             Set-ItemProperty -Path "HKCU:\System\GameConfigStore" -Name "GameDVR_EFSEFeatureFlags" -Type DWord -Value 0
             Set-ItemProperty -Path "HKCU:\System\GameConfigStore" -Name "GameDVR_Enabled" -Type DWord -Value 0
             Set-ItemProperty -Path "HKCU:\System\GameConfigStore" -Name "GameDVR_FSEBehavior" -Type DWord -Value 2
+            If (!(Test-Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\GameDVR")) {
+                New-Item -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\GameDVR" -Force
+            }
             Set-ItemProperty -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows\GameDVR" -Name "AllowGameDVR" -Type DWord -Value 0
 
             #Disabling Gamebar Presence Writer, which causes stutter in games
@@ -844,9 +839,11 @@ $WPFtweaksbutton.Add_Click({
             $WPFMiscTweaksRightClickMenu.IsChecked = $false
         }
         If ( $WPFEssTweaksOO.IsChecked -eq $true ) {
-            Write-Host "Running O&O Shutup with Recommended Settings"
-            curl.exe -ss "https://raw.githubusercontent.com/ChrisTitusTech/win10script/master/ooshutup10.cfg" -o ooshutup10.cfg
-            curl.exe -ss "https://dl5.oo-software.com/files/ooshutup10/OOSU10.exe" -o OOSU10.exe
+            If (!(Test-Path .\ooshutup10.cfg)) {
+                Write-Host "Running O&O Shutup with Recommended Settings"
+                curl.exe -ss "https://raw.githubusercontent.com/ChrisTitusTech/win10script/master/ooshutup10.cfg" -o ooshutup10.cfg
+                curl.exe -ss "https://dl5.oo-software.com/files/ooshutup10/OOSU10.exe" -o OOSU10.exe
+            }
             ./OOSU10.exe ooshutup10.cfg /quiet
             $WPFEssTweaksOO.IsChecked = $false
         }
@@ -952,7 +949,7 @@ $WPFtweaksbutton.Add_Click({
                 # -ErrorAction SilentlyContinue is so it doesn't write an error to stdout if a service doesn't exist
         
                 Write-Host "Setting $service StartupType to Manual"
-                Get-Service -Name $service -ErrorAction SilentlyContinue | Set-Service -StartupType Manual
+                Get-Service -Name $service -ErrorAction SilentlyContinue | Set-Service -StartupType Manual -ErrorAction SilentlyContinue
             }
             $WPFEssTweaksServices.IsChecked = $false
         }
@@ -1065,17 +1062,17 @@ $WPFtweaksbutton.Add_Click({
             ## Performance Tweaks and More Telemetry
             Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\DriverSearching" -Name "SearchOrderConfig" -Type DWord -Value 0
             Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Multimedia\SystemProfile" -Name "SystemResponsiveness" -Type DWord -Value 0
-            Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Control" -Name "WaitToKillServiceTimeout" -Type DWord -Value 2000
             Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "MenuShowDelay" -Type DWord -Value 1
-            Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "WaitToKillAppTimeout" -Type DWord -Value 5000
-            Remove-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "HungAppTimeout" -ErrorAction SilentlyContinue
-            # Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "HungAppTimeout" -Type DWord -Value 4000 # Note: This caused flickering
             Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "AutoEndTasks" -Type DWord -Value 1
-            Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "LowLevelHooksTimeout" -Type DWord -Value 1000
-            Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "WaitToKillServiceTimeout" -Type DWord -Value 2000
             Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\Memory Management" -Name "ClearPageFileAtShutdown" -Type DWord -Value 0
             Set-ItemProperty -Path "HKCU:\Control Panel\Mouse" -Name "MouseHoverTime" -Type DWord -Value 10
-
+            
+            ## Timeout Tweaks cause flickering on Windows now
+            Remove-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "WaitToKillAppTimeout" -ErrorAction SilentlyContinue
+            Remove-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "HungAppTimeout" -ErrorAction SilentlyContinue
+            Remove-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Control" -Name "WaitToKillServiceTimeout" -ErrorAction SilentlyContinue
+            Remove-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "LowLevelHooksTimeout" -ErrorAction SilentlyContinue
+            Remove-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "WaitToKillServiceTimeout" -ErrorAction SilentlyContinue
 
             # Network Tweaks
             Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\LanmanServer\Parameters" -Name "IRPStackSize" -Type DWord -Value 20
@@ -1197,7 +1194,7 @@ $WPFtweaksbutton.Add_Click({
         }
         If ( $WPFEssTweaksRemoveEdge.IsChecked -eq $true ) {
             Write-Host "Removing Microsoft Edge..."
-            Invoke-WebRequest -useb https://raw.githubusercontent.com/ChrisTitusTech/winutil/main/Edge_Removal.bat | Invoke-Expression
+            Invoke-WebRequest -useb https://raw.githubusercontent.com/ChrisTitusTech/winutil/$BranchToUse/Edge_Removal.bat | Invoke-Expression
             $WPFEssTweaksRemoveEdge.IsChecked = $false
         }
         If ( $WPFEssTweaksDeBloat.IsChecked -eq $true ) {
